Moved Living->getDrops to Entity->getDrops
